+++ bombshell/__init__.py
@@ -1,5 +1,6 @@
 from .core import exec, Process
 from .resources import ResourceData
 from .results import CompletedProcess, PipelineError
+from .spin import spin
 
-__all__ = ["exec", "Process", "ResourceData", "CompletedProcess", "PipelineError"]
+__all__ = ["exec", "Process", "ResourceData", "CompletedProcess", "PipelineError", "spin"]